Added streaming support for Android
Added a new event that can be used to subscribe to audio data events while recording. We try to send an event every 10ms, but Android may force a longer frame time. We don't resample the audio data, but we provide the true sample rate so callers can adapt as needed.